requirements.txt
aiogram==3.21.0
fastapi==0.116.1
uvicorn==0.35.0
SQLAlchemy==2.0.43
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.12.0
python-dotenv==1.1.1
pydantic==2.11.7
yookassa==3.0.1
colorlog==6.9.0
watchfiles==1.1.0
yt-dlp==2024.10.22
moviepy==2.2.1
